refactor(cli): move the web file commands to the neutral root
Move the Web command domain to the neutral command root and update its internal command export without changing public CLI or API behavior.
